+++ src/components/Card/PetCard/PetDetail.tsx
